Take the lint gates to zero
Allow deref_addrof at each port's crate root — clippy's fix does not
compile under edition 2024 — and fix the 67 real findings behind it, so
every architecture baseline is zero and the host lint is clean.

+++ kernel/kernel-riscv64/src/discovery.rs
@@ -312,7 +312,7 @@ pub(crate) fn boot_memory_map(
         // The device tree itself, reclaimable once discovery has consumed
         // it — which has not happened yet, so it stays reserved for now.
         MemoryRegion {
-            base: PhysAddr::new(dtb as u64 - DIRECT_MAP_BASE),
+            base: PhysAddr::new(dtb - DIRECT_MAP_BASE),
             len: tree.len() as u64,
             kind: MemoryKind::BootloaderReclaimable,
         },
